listener/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-servicebus  # For quick_enqueue.py testing script
python-dotenv  # For quick_enqueue.py testing script

psycopg2-binary
